refactor!: roll ProcessMasterlist class into ProcessData as internal
